requirements.txt
django
gunicorn
robotframework
selenium
robotframework-selenium2library
